+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.4.7
+
+### Patch Changes
+
+- d5c378b6: fix: reduce bundle size by remove runtime module
+- 77155bab: feat: remove runtime code when loaders is not export
+- Updated dependencies [d5c378b6]
+- Updated dependencies [77155bab]
+  - @ice/runtime@1.4.5
+
 ## 3.4.6
 
 ### Patch Changes